move cli to lib/cli
