+++ src/commands/admin/generate-api-keys/command.ts
@@ -1,6 +1,6 @@
 import { buildCommand } from '@stricli/core';
-import { createTranscendUrlParameter } from '@/cli/common-parameters';
-import { uuidParser } from '@/cli/parsers';
+import { createTranscendUrlParameter } from '@/lib/cli/common-parameters';
+import { uuidParser } from '@/lib/cli/parsers';
 
 export const generateApiKeysCommand = buildCommand({
   loader: async () => {
