+++ src/commands/consent/upload-cookies-from-csv/command.ts
@@ -3,7 +3,7 @@ import { ConsentTrackerStatus, ScopeName } from '@transcend-io/privacy-types';
 import {
   createAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
+} from '@/lib/cli/common-parameters';
 
 export const uploadCookiesFromCsvCommand = buildCommand({
   loader: async () => {
