+++ src/commands/inventory/derive-data-silos-from-data-flows-cross-instance/command.ts
@@ -2,7 +2,7 @@ import { buildCommand } from '@stricli/core';
 import {
   createAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
+} from '@/lib/cli/common-parameters';
 
 export const deriveDataSilosFromDataFlowsCrossInstanceCommand = buildCommand({
   loader: async () => {
