+++ src/commands/request/cron/mark-identifiers-completed/command.ts
@@ -4,8 +4,8 @@ import {
   createAuthParameter,
   createSombraAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
-import { uuidParser } from '@/cli/parsers';
+} from '@/lib/cli/common-parameters';
+import { uuidParser } from '@/lib/cli/parsers';
 
 export const markIdentifiersCompletedCommand = buildCommand({
   loader: async () => {
