+++ src/commands/request/download-files/command.ts
@@ -4,8 +4,8 @@ import {
   createAuthParameter,
   createSombraAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
-import { dateParser } from '@/cli/parsers';
+} from '@/lib/cli/common-parameters';
+import { dateParser } from '@/lib/cli/parsers';
 
 export const downloadFilesCommand = buildCommand({
   loader: async () => {
