+++ src/commands/request/notify-additional-time/command.ts
@@ -3,8 +3,8 @@ import { RequestAction, ScopeName } from '@transcend-io/privacy-types';
 import {
   createAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
-import { dateParser } from '@/cli/parsers';
+} from '@/lib/cli/common-parameters';
+import { dateParser } from '@/lib/cli/parsers';
 
 export const notifyAdditionalTimeCommand = buildCommand({
   loader: async () => {
