+++ src/commands/request/preflight/pull-identifiers/command.ts
@@ -5,7 +5,7 @@ import {
   createAuthParameter,
   createSombraAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
+} from '@/lib/cli/common-parameters';
 
 export const pullIdentifiersCommand = buildCommand({
   loader: async () => {
