+++ src/commands/request/reject-unverified-identifiers/command.ts
@@ -3,7 +3,7 @@ import { RequestAction, ScopeName } from '@transcend-io/privacy-types';
 import {
   createAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
+} from '@/lib/cli/common-parameters';
 
 export const rejectUnverifiedIdentifiersCommand = buildCommand({
   loader: async () => {
