+++ src/commands/request/system/retry-request-data-silos/command.ts
@@ -3,8 +3,8 @@ import { RequestAction, ScopeName } from '@transcend-io/privacy-types';
 import {
   createAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
-import { uuidParser } from '@/cli/parsers';
+} from '@/lib/cli/common-parameters';
+import { uuidParser } from '@/lib/cli/parsers';
 
 export const retryRequestDataSilosCommand = buildCommand({
   loader: async () => {
