+++ src/commands/request/upload/command.ts
@@ -4,7 +4,7 @@ import {
   createAuthParameter,
   createSombraAuthParameter,
   createTranscendUrlParameter,
-} from '@/cli/common-parameters';
+} from '@/lib/cli/common-parameters';
 
 export const uploadCommand = buildCommand({
   loader: async () => {
