+++ src/lib/cli/legacy-commands.ts
@@ -3,7 +3,7 @@ import {
   type Application,
   type CommandContext,
 } from '@stricli/core';
-import { app } from '../app';
+import { app } from '../../app';
 import { logger } from '@/logger';
 
 // Maps legacy (<7.0.0) command names to their modern command structure
